Split MutableHistoryImpl and return an optimized MutableEventLog when calling `mutableEventLogOf()`
