+++ echo-core/src/commonMain/kotlin/io/github/alexandrepiveteau/echo/core/log/Builders.kt
@@ -38,7 +38,7 @@ fun <T> mutableHistoryOf(
     }
 
 /** Creates a new [MutableEventLog], with no aggregate. */
-fun mutableEventLogOf(): MutableEventLog = mutableHistoryOf(NoModel, NoProjection)
+fun mutableEventLogOf(): MutableEventLog = mutableEventLogOf(*emptyArray())
 
 /**
  * Creates a new [MutableEventLog], with no aggregate.
@@ -47,30 +47,9 @@ fun mutableEventLogOf(): MutableEventLog = mutableHistoryOf(NoModel, NoProjectio
  */
 fun mutableEventLogOf(
     vararg events: Pair<EventIdentifier, ByteArray>,
-): MutableEventLog = mutableHistoryOf(NoModel, NoProjection, *events)
-
-// An object which represents the absence of an aggregated model.
-private object NoModel
-
-// An object which represents the absence of an aggregating projection.
-private object NoProjection : MutableProjection<NoModel> {
-
-  override fun ChangeScope.forward(
-      model: NoModel,
-      identifier: EventIdentifier,
-      data: ByteArray,
-      from: Int,
-      until: Int,
-  ): NoModel = NoModel
-
-  override fun backward(
-      model: NoModel,
-      identifier: EventIdentifier,
-      data: ByteArray,
-      from: Int,
-      until: Int,
-      changeData: ByteArray,
-      changeFrom: Int,
-      changeUntil: Int,
-  ): NoModel = NoModel
-}
+): MutableEventLog =
+    MutableEventLogImpl().apply {
+      for ((id, body) in events) {
+        insert(id.seqno, id.site, body)
+      }
+    }
